bugfix for flying annotations
when viewing annotations inside a building with the firstPersonView active, it appears that the annotations are somehow mirrored at the user's viewpoint. So you sometimes have 2 annotation, one of them flying kind of randomly in the air... Checking if tempVec4b[2]<0 and in that case setting one of the coordinates outside the user's screen boundaries fixes that.