zoom on empty canvas leads to infinity zoomRate
In an initial setup with an emtpy canvas, using the mousewheel to scroll leads to infinity values while calculating the zoomRate using the scenes aabb. 
A solution would be returning a zoomRate of 0, if theres is nothing loaded into the scene. 
Another approach would be to calculate the aabb of all loaded models in the scene not using the scenes aabb by itself.